+++ core/moengage_flutter_platform_interface/lib/src/internal/method_channel_moengage_flutter.dart
@@ -279,4 +279,10 @@ class MethodChannelMoEngageFlutter extends MoEngageFlutterPlatform {
     payload[keyData] = {keyUpdatePushPermissionCount: requestCount};
     return jsonEncode(payload);
   }
+
+  @override
+  void showNudge(MoEngageNudgePosition position, String appId) {
+    _methodChannel.invokeMethod(methodNameShowNudge,
+        jsonEncode(getShowNudgeJsonPayload(position, appId)));
+  }
 }