+++ example/android/app/src/main/java/com/moengage/sampleapp/CustomPushListener.kt
@@ -4,23 +4,20 @@ import android.app.Activity
 import android.content.Context
 import android.os.Bundle
 import com.moengage.core.internal.logger.Logger
+import com.moengage.core.model.AccountMeta
 import com.moengage.plugin.base.push.PluginPushCallback
 
 /**
  * @author Umang Chamaria
  * Date: 2020/12/06
  */
-class CustomPushListener : PluginPushCallback() {
+class CustomPushListener(accountMeta: AccountMeta) : PluginPushCallback(accountMeta) {
 
     private val tag = "CustomPushListener"
-    override fun handleCustomAction(context: Context, payload: String) {
-        super.handleCustomAction(context, payload)
-        Logger.print { "$tag handleCustomAction() : " }
-    }
 
-    override fun onNotificationClick(activity: Activity, payload: Bundle) {
-        super.onNotificationClick(activity, payload)
+    override fun onNotificationClick(activity: Activity, payload: Bundle): Boolean {
         Logger.print { "$tag onNotificationClick() : " }
+        return super.onNotificationClick(activity, payload)
     }
 
 }