+++ cards/moengage_cards_android/CHANGELOG.md
@@ -1,5 +1,12 @@
 # MoEngage Cards Android Plugin
 
+# Next Release
+
+## Next Version
+- Support for `moe-android-sdk` version `13.00.00` and above
+- Add support for AGP `8.2.2` and above
+- Support for `cards-core` version `2.0.0` and above
+
 # 07-12-2023
 
 # 1.1.0