+++ geofence/moengage_geofence_android/CHANGELOG.md
@@ -1,5 +1,12 @@
 # MoEngage Geofence Android Plugin
 
+# Next Release
+
+## Next Version
+- Support for `moe-android-sdk` version `13.00.00` and above
+- Add support for AGP `8.2.2` and above
+- Support for `geofence` version `4.0.0` and above
+
 # 07-12-2023
 
 # 1.1.0