+++ inbox/moengage_inbox_android/CHANGELOG.md
@@ -1,5 +1,12 @@
 # MoEngage Inbox Android Plugin
 
+# Next Release
+
+## Next Version
+- Support for `moe-android-sdk` version `13.00.00` and above
+- Add support for AGP `8.2.2` and above
+- Support for `inbox-core` version `3.0.0` and above
+
 # 07-12-2023
 
 # 1.1.0